move consent retrieval to amp layer add jwplayer fixture
