+++ extensions/amp-jwplayer/1.0/amp-jwplayer.js
@@ -1,14 +1,57 @@
+import {dict} from '#core/types/object';
+
 import {isExperimentOn} from '#experiments';
 
 import {userAssert} from '#utils/log';
 
 import {BaseElement} from './base-element';
 
+import {
+  getConsentMetadata,
+  getConsentPolicyInfo,
+  getConsentPolicyState,
+} from '../../../src/consent';
+
 /** @const {string} */
 const TAG = 'amp-jwplayer';
 
 /** @implements {../../../src/video-interface.VideoInterface} */
 class AmpJwplayer extends BaseElement {
+  /** @override */
+  init() {
+    const consentPolicy = this.getConsentPolicy();
+    if (consentPolicy) {
+      this.getConsentInfo().then((consentInfo) => {
+        const policyState = consentInfo[0];
+        const policyInfo = consentInfo[1];
+        const policyMetadata = consentInfo[2];
+        this.mutateProps(
+          dict({
+            'consentParams': {
+              'policyState': policyState,
+              'policyInfo': policyInfo,
+              'policyMetadata': policyMetadata,
+            },
+          })
+        );
+      });
+    }
+
+    return super.init();
+  }
+
+  /**
+   * @param {string} policy
+   * @return {Promise}
+   */
+  getConsentInfo(policy) {
+    return Promise.all([
+      getConsentPolicyState(this.element, policy),
+      getConsentPolicyInfo(this.element, policy),
+      getConsentMetadata(this.element, policy),
+    ]);
+  }
+
   /** @override */
   isLayoutSupported(layout) {
     userAssert(
