+++ test/fixtures/e2e/bento/jwplayer.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <meta charset="utf-8" />
+  <title>Bento Jwplayer</title>
+  <script type="module" async src="https://cdn.ampproject.org/bento.mjs"></script>
+
+  <script async src="https://cdn.ampproject.org/v0/bento-jwplayer-1.0.js"></script>
+  <link rel="stylesheet" href="https://cdn.ampproject.org/v0/bento-jwplayer-1.0.css">
+  <meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1" />
+
+</head>
+
+<body>
+
+  <bento-jwplayer
+    id="jwplayer"
+    data-player-id="BjcwyK37"
+    data-media-id="CtaIzmFs"
+    style="width: 480px; height: 270px"
+  ></bento-jwplayer>
+
+
+</body>
+
+</html>
